APP: 微调优化

+++ app/src/main/java/com/kangear/bodycompositionanalyzer/HeightActivity.java
@@ -75,7 +75,7 @@ public class HeightActivity extends Com2Activity {
     public void onNextButtonClick() {
         super.onNextButtonClick();
         Log.i(TAG, "onNextButtonClick");
-        WelcomeActivity.getPerson().setHeight(Integer.valueOf(mEditText.getText().toString()));
+        WelcomeActivity.getPerson().setHeight(Float.valueOf(mEditText.getText().toString()));
         WelcomeActivity.doTest(this);
         finish();
     }
